parse_text: return if next is UTF8_INVALID
In the underscore case, the next character is retrieved to check
whether the underscore is at a word break. However, if this character
is UTF8_INVALID, the call to parser_pushch will be a noop. This
results in the loop continuing on further than it should. This just
adds a check to see if next is UTF8_INVALID and returns if it is.

Signed-off-by: Brian Ashworth <bosrsf04@gmail.com>